Split out getDomainProductRanking
