+++ client/lib/products-values/get-domain-product-ranking.js
@@ -1,0 +1,18 @@
+/**
+ * Internal dependencies
+ */
+import { assertValidProduct } from 'lib/products-values/utils/assert-valid-product';
+import { formatProduct } from 'lib/products-values/format-product';
+import { isDomainRegistration } from 'lib/products-values/is-domain-registration';
+import { isDomainMapping } from 'lib/products-values/is-domain-mapping';
+
+export function getDomainProductRanking( product ) {
+	product = formatProduct( product );
+	assertValidProduct( product );
+
+	if ( isDomainRegistration( product ) ) {
+		return 0;
+	} else if ( isDomainMapping( product ) ) {
+		return 1;
+	}
+}
